+++ src/logger.js
@@ -21,7 +21,5 @@
 
 export default class NullLogger {
     info(msg: string): void {}
-    debug(msg: string): void {}
-    warn(msg: string): void {}
     error(msg: string): void {}
 }
